fix: prisma rm db script.
