+++ app/web-app/src/mocks/remove.ts
@@ -1,0 +1,29 @@
+import { PrismaClient } from '@prisma/client';
+
+const db = new PrismaClient();
+
+async function main() {
+    const tableNames = await db.$queryRaw<Array<{ tablename: string }>>`
+            SELECT tablename FROM pg_tables WHERE schemaname='public'
+        `;
+
+    for (const { tablename } of tableNames) {
+        if (tablename !== '_prisma_migrations') {
+            await db.$executeRawUnsafe(
+                `TRUNCATE TABLE "${tablename}" CASCADE;`
+            );
+        }
+    }
+    return { message: 'DB cleared successfully.' };
+}
+
+main()
+    .then((res) => {
+        console.log(res?.message);
+    })
+    .catch((e) => {
+        console.error(e);
+    })
+    .finally(async () => {
+        await db.$disconnect();
+    });
